requirements.txt
Flask==2.3.3
openai-whisper==20230314
torch==2.0.1
torchaudio==2.0.2
numpy==1.24.4
gunicorn==21.2.0

